Generated by Spark: Hi Copilot, the goal of this project is. The goal of this project  is to create a table with advanced UX behaviour such as sticky behaviour.
In depth it should have the following features:

* Standard table behavior such as showing data, either from in memory or from a http resource
* Have a paginator which is customizable in how many items per page should be shown. Via in code configuration these values can be choosen or fixed to a certain amount. In the case of only one page we should allow to configure whether or not we want to show the pager.
* Have the possibility to expand rows to show a detailed view of the item (custom template)
* Have the possibility to sort rows (justh the UI/ UX, it should propagate an event to which field is sorted and in which matter)
* Have the possibility for dark/ light theme with Tailwind support
* Have the possibility of frozen columns (which are in one html table element, not outside!)
* HAve the possibility to determine how columns in width should behave, either overflow ellipsed with min/max value as going multiline or no max width at all
* A horizontal scrollbar should occur when the width is too big
* The vertical height follows the parent alement unless a fixed width is given to the table
* When scrolling down, the header becomes sticky and follows along with it

Technically, it should be done in Angular using the best practices and template driven for the rows and headers. Every action should event driven. Loading data and the actual paging is just a presentation concern for the table. Via event propagation it is requested in the parent component to change the data for another or sorted page.

An idea of how the table should behave can be get from Mailchimp. They have such a sticky header. For some reason they have two theads as well.